首页与达人申请优化

+++ TalentService/TalentService/Classes/Activity(活动)/View/ActivityDetail(活动详情)/ZZActivityBottomToolBar.m
@@ -68,7 +68,7 @@
     [super layoutSubviews];
     
     // 设置按钮的frame
-    int btnCount = self.btns.count;
+    NSInteger btnCount = self.btns.count;
     CGFloat btnW = self.width / btnCount;
     CGFloat btnH = self.height;
     for (int i = 0; i<btnCount; i++) {
